requirements_dev.txt
# Testing
pytest>=6.2.5,<7.0
pytest-cov>=2.10.1,<3.0
pytest-mock>=3.6.1,<4.0
pytest-xdist>=2.5.0,<3.0

# Development
ipdb>=0.13.9,<1.0
icecream>=2.1.1,<3.0
elasticsearch
sentence-transformers
streamlit
gitpython
pathlib